LPD-68061 Split Salesforce wizard into pieces
