+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/settings/pages/data-source/salesforce/utis.ts
@@ -1,0 +1,9 @@
+export function updateSearchParams(history, key: string, value: any) {
+	const params = new URLSearchParams(window.location.search);
+	params.set(key, String(value));
+
+	history.push({
+		pathname: window.location.pathname,
+		search: params.toString()
+	});
+}
